+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Upcoming event order is now correct
 - ``BookingController@edit()`` had one ``flashMessage()`` that did not show correctly
+- Airports now use a ``orderBy('ICAO')``
 
 ## [v0.7.0b](https://gitlab.com/daveroverts/Book-me-a-cookie/compare/v0.7.0a...v0.7.0b) - 2018-11-25